% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_reference_data.R
\name{as.bmseList}
\alias{as.bmseList}
\title{Verifying an entry exists for a given CAS number and solvent type}
\usage{
as.bmseList(
  casno_list,
  return_metabs = "exact_match",
  solvent_type = NULL,
  ph = NULL,
  instrument_strength
)
}
\arguments{
\item{casno_list}{list of CAS registry numbers written as characters}

\item{return_metabs}{string, must be one of "exact_match", "nearest_match", or "all". Defaults to "exact_match". If "exact_match", returns metabolites that exactly match specified experimental conditions (solvent type, pH, and instrument strength). If "nearest_match", returns nearest match metabolites, sorted by Euclidean distance. If "all", returns all entries corresponding to supplied CAS numbers, experimental conditions ignored.}

\item{solvent_type}{choose from available solvents 'D2O', 'H2O', ...}

\item{ph}{numeric value specifying pH of experimental conditions}

\item{instrument_strength}{numeric value specifying experimental instrument strength}
}
\value{
a list of BMSE database entry numbers
}
\description{
Verifying an entry exists for a given CAS number and solvent type
}
\examples{
# test <- as.bmseList(list("56-41-7","2613-02-7"), solvent_type = 'D2O', ph = 7.4)

}
