% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_reference_data.R
\name{as.bmseListFromName}
\alias{as.bmseListFromName}
\title{Verifying an entry exists for a given metabolite name}
\usage{
as.bmseListFromName(
  name_list,
  return_metabs = "exact_match",
  solvent_type = NULL,
  ph = NULL,
  instrument_strength
)
}
\arguments{
\item{name_list}{list of metabolite names}

\item{return_metabs}{}

\item{solvent_type}{choose from available solvents 'D2O', 'H2O', ...}

\item{ph}{numerical value specifying pH of the experimental conditions}

\item{instrument_strength}{numeric value specifying experimental instrument strength}
}
\value{
a list of BMSE database entry numbers
}
\description{
Verifying an entry exists for a given metabolite name
}
\examples{
# test <- as.bmseListFromName(list("ATP","Maltose" "Oxalate"), solvent_type = 'D2O', ph = 7.4)

}
