% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bmse_associations}
\alias{bmse_associations}
\title{BMSE Associations}
\format{
A data frame with 2224 rows and 9 columns:
\describe{
  \item{Entry_ID}{BMSE identifier}
  \item{CASno}{CAS number}
  \item{Field_strength}{}
  \item{Solute}{}
  \item{Solvent}{Experimental conditions}
  \item{Reference}{}
  \item{pH}{Experimental conditions}
  \item{Temperature}{Experimental conditions}
  \item{Concentration}{Experimental Conditions}
}
}
\usage{
bmse_associations
}
\description{
A dataset containing metabolite data
}
\keyword{datasets}
