% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{e_data}
\alias{e_data}
\title{Example e_data}
\format{
A data frame with 262144 rows and 5 columns:
\describe{
  \item{PPM}{Observed PPMs in NMR spectra}
  \item{AA_1x_a_20201209.17.fid.1}{Sample data at each PPM}
  \item{AA_1x_a_20201209.18.fid.2}{Sample data at each PPM}
  \item{AA_1x_a_20201209.19.fid.3}{Sample data at each PPM}
  \item{AA_1x_a_20201209.20.fid.4}{Sample data at each PPM}
}
}
\usage{
e_data
}
\description{
Copyright (C) 2022 Battelle Memorial Institute

 This program is free software; you can redistribute it and/or modify
 it under the terms of the GNU General Public License as published by
 the Free Software Foundation; either version 2 of the License, or
 (at your option) any later version.

 This program is distributed in the hope that it will be useful,
 but WITHOUT ANY WARRANTY; without even the implied warranty of
 MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
 GNU General Public License for more details.

 You should have received a copy of the GNU General Public License along
 with this program; if not, write to the Free Software Foundation, Inc.,
 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.

A dataset containing NMR spectral data over 4 samples
}
\keyword{datasets}
