% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edit_reference_data.R
\name{edit_reference_data}
\alias{edit_reference_data}
\title{Edit Reference Data}
\usage{
edit_reference_data(
  roi_df,
  metabolite_name,
  ppm,
  tolerance = 0.05,
  parameter_name,
  parameter_value
)
}
\arguments{
\item{roi_df}{data frame returned from roi_ref_export() that contains the rDolphin reference data}

\item{metabolite_name}{character vector with one or more entries}

\item{ppm}{numeric vector with one or more entries; chemical shift of the parameter to be changed}

\item{tolerance}{chemical shift tolerance to search for a peak within; default is 0.05}

\item{parameter_name}{character vector with one or more entries; parameters to be edited}

\item{parameter_value}{character vector with one or more entries; the new value a parameter will be set to with the specified identifiers}
}
\value{
an edited data.frame formatted for rDolphin use
}
\description{
Edit Reference Data
}
