% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_metabs}
\alias{example_metabs}
\title{Example metabolites}
\format{
A data frame with 10 rows and 12 columns:
\describe{
  \item{ROI left edge (ppm)}{Left boundary of region of interest}
  \item{ROI right edge (ppm)}{Right boundary of region of interest}
  \item{Quantification Mode}{}
  \item{Metabolite}{}
  \item{Quantification Signal}{}
  \item{Chemical shift(ppm)}{}
  \item{Chemical shift tolerance (ppm)}{}
  \item{Half bandwidth (Hz)}{}
  \item{Multiplicity}{}
  \item{J coupling (Hz)}{}
  \item{Roof effect}{}
  \item{HMDB_code}{}
}
}
\usage{
example_metabs
}
\description{
Copyright (C) 2022 Battelle Memorial Institute

 This program is free software; you can redistribute it and/or modify
 it under the terms of the GNU General Public License as published by
 the Free Software Foundation; either version 2 of the License, or
 (at your option) any later version.

 This program is distributed in the hope that it will be useful,
 but WITHOUT ANY WARRANTY; without even the implied warranty of
 MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
 GNU General Public License for more details.

 You should have received a copy of the GNU General Public License along
 with this program; if not, write to the Free Software Foundation, Inc.,
 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.

A dataset containing example metabolite data
}
\keyword{datasets}
