% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_ppm.R
\name{filter_ppm}
\alias{filter_ppm}
\title{PPM Filtering}
\usage{
filter_ppm(ppmData, range)
}
\arguments{
\item{ppmData}{a list object from containing the data tables `e_data` and `f_data`, created using the function \code{\link{as.ppmData}}}

\item{range}{a list specifying the minimum and maximum value(s) in the range(s) to be filtered. Multiple ranges may be specified. Ranges are inclusive.}
}
\value{
`ppmData` object with 'filters' attribute including range and `data.frame` with removed rows
}
\description{
Filters out specified PPM ranges from the dataset
}
\examples{
\dontrun{
filter_ppm(mydata, range = list(min = 2, max = 3))
filter_ppm(mydata, range = list(min = c(1,3), max = c(2,4)))
}

}
\author{
Allison Thompson, Natalie Winans
}
