% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rD_fitting_prep.R
\name{fitting_prep}
\alias{fitting_prep}
\title{rDolphin: fitting_prep}
\usage{
fitting_prep(
  Xdata,
  Ydata,
  initial_fit_parameters,
  program_parameters,
  created_baseline
)
}
\arguments{
\item{Xdata}{}

\item{Ydata}{}

\item{initial_fit_parameters}{}

\item{program_parameters}{}

\item{created_baseline}{}
}
\description{
rDolphin: fitting_prep
}
\details{
This function was forked directly from rDolphin (https://github.com/danielcanueto/rDolphin/tree/master/R). This is function is used to collate 
the data necessary for the downstream fitting algorithm and associated functions. Modifications have been made to the original function to allow for
the fitting of double-doublets. Generally speaking, this involves the addition of new fields (a second J-coupling field and second
roof effect field) containing parameters necessary to fit this new pattern.
}
