% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rD_fitting_prep_2.R
\name{fitting_prep_2}
\alias{fitting_prep_2}
\title{rDolphin: fitting_prep_2}
\usage{
fitting_prep_2(
  Xdata,
  Ydata,
  initial_fit_parameters,
  program_parameters,
  created_baseline,
  max_shift,
  min_shift,
  max_intensity,
  min_intensity,
  max_width,
  min_width,
  spectrum_index,
  signal_index
)
}
\arguments{
\item{Xdata}{}

\item{Ydata}{}

\item{initial_fit_parameters}{}

\item{program_parameters}{}

\item{created_baseline}{}

\item{max_shift}{}

\item{min_shift}{}

\item{max_intensity}{}

\item{min_intensity}{}

\item{max_width}{}

\item{min_width}{}

\item{spectrum_index}{}

\item{signal_index}{}
}
\description{
rDolphin: fitting_prep_2
}
\details{
This function was forked directly from rDolphin (https://github.com/danielcanueto/rDolphin/tree/master/R). This is function is used to collate 
the data necessary for the downstream fitting algorithm and associated functions, assuming that profiling has occurred at least once before.
}
