% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_plotting.R
\name{format_plotting}
\alias{format_plotting}
\title{Format results for plotting}
\usage{
format_plotting(profiling_data, signals_to_plot = NULL)
}
\arguments{
\item{profiling_data}{Resultant object from \code{\link{run_rDolphin}}}

\item{signals_to_plot}{Vector of indexes of signals in ROI data to plot. By default, NA and all figures are outputted.}
}
\value{
Data object for Trelliscope Plotting
}
\description{
Format results for plotting
}
\author{
Allison Thompson
}
