% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_reference_data.R
\name{get_spectra_data}
\alias{get_spectra_data}
\title{Pulls spectra character data for a given entry id from a locally maintained respository of JSON formated BMRM-STAR files}
\usage{
get_spectra_data(ID_list)
}
\arguments{
\item{ID_list}{list of BMRB IDs; For metabolomics entries it is just a number with BMRB prefix (example: 'bmse000028')}
}
\value{
R data frame that contains  Spectral_transition_char data for a given entry ID
}
\description{
Copyright (C) 2022 Battelle Memorial Institute

 This program is free software; you can redistribute it and/or modify
 it under the terms of the GNU General Public License as published by
 the Free Software Foundation; either version 2 of the License, or
 (at your option) any later version.

 This program is distributed in the hope that it will be useful,
 but WITHOUT ANY WARRANTY; without even the implied warranty of
 MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
 GNU General Public License for more details.

 You should have received a copy of the GNU General Public License along
 with this program; if not, write to the Free Software Foundation, Inc.,
 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
}
\examples{
# df <- get_spectra_data("path/to/json/", as.bmseList(list("56-41-7","2613-02-7"), 400, 'D2O'))
# Downloads data from BMRB metabolomics database
}
\seealso{
\code{\link{show_file}}
}
