% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_utils.R
\name{getmode}
\alias{getmode}
\title{Return the mode of a vector}
\usage{
getmode(x, ...)
}
\arguments{
\item{x}{A vector}

\item{...}{Additional arguments to pass to \code{table()}}
}
\value{
A character-valued scalar.
}
\description{
Copyright (C) 2022 Battelle Memorial Institute

 This program is free software; you can redistribute it and/or modify
 it under the terms of the GNU General Public License as published by
 the Free Software Foundation; either version 2 of the License, or
 (at your option) any later version.

 This program is distributed in the hope that it will be useful,
 but WITHOUT ANY WARRANTY; without even the implied warranty of
 MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
 GNU General Public License for more details.

 You should have received a copy of the GNU General Public License along
 with this program; if not, write to the Free Software Foundation, Inc.,
 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
}
\details{
This function determines the most common value in a vector. In the event of a tie, this function returns
the value that is first as ordered by \code{table()}. Whether the vector is numeric or character, this
function returns a character-valued scalar.
}
