% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rD_integration.R
\name{integration}
\alias{integration}
\title{rDolphin: integration}
\usage{
integration(clean_fit, Xdata, Ydata, buck_step, interface = "F")
}
\arguments{
\item{clean_fit}{}

\item{Xdata}{}

\item{Ydata}{}

\item{buck_step}{}

\item{interface}{}
}
\description{
rDolphin: integration
}
\details{
This function was forked directly from rDolphin (https://github.com/danielcanueto/rDolphin/tree/master/R). This function performs quantification through integration.
}
