% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{load_file}
\alias{load_file}
\title{Load experimental NMR data, experimental NMR metadata, or target metabolite data}
\usage{
load_file(path, dataset)
}
\arguments{
\item{path}{The filepath of the data of interest.}

\item{dataset}{The type of data to be uploaded. Users may specify either "experiment", "experiment_metadata", 
or "metabolites" depending on whether experimental data, experimental metadata, or target metabolite data are being uploaded, respectively.}
}
\value{
A dataframe containing the uploaded data.
}
\description{
Load experimental NMR data, experimental NMR metadata, or target metabolite data
}
\details{
If users are uploading experimental data, it is assumed that the data are contained within a .csv file and 
that these data are organized such that the first column contains the PPM values of the experimental spectra and
each subsequent column contains intensity values for a given sample at the corresponding PPM value. 

If users are uploading experimental metadata, it is assumed that the data are contained within a .csv file, that the first column 
of these data contain the sample name(s), and the second column contains an experimental label. 

Last, uploaded target metabolite data is assumed to be contained within an .xlsx file that contains a column of CAS numbers 
corresponding to the metabolites of interest.
}
