% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_nmRapp_server.R
\name{nmRapp_server}
\alias{nmRapp_server}
\title{Main server function of nmRapp}
\usage{
nmRapp_server(input, output, session)
}
\arguments{
\item{input}{A list-like object containing all input data sent from the browser. Elements of this list are
named according to their input ID.}

\item{output}{A list-like object containing all output data, Elements of this list are named according to
their output ID.}

\item{session}{A new instance of the shiny app.}
}
\description{
Main server function of nmRapp
}
\details{
This main server function references each of the defined server module components that handle various
aspects of the NMR data processing pipeline specific to targeted metabolite profiling. Refer to the documentation
of each referenced module for details on its specific function. A list of referenced modules is provided below:
1) xpmt_data_uploadServer()
2) xpmt_data_vizServer()
3) xpmt_data_modifyServer()
4) ref_data_uploadServer()
5) ref_data_editingServer

Outside of calls to the above modules, this main server function handles the dynamic generation of UI elements that
are external to each module (e.g. global tabs) and data processing dependent on output from multiple independent modules.
}
