% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_nmRapp_ui.R
\name{nmRapp_ui}
\alias{nmRapp_ui}
\title{Main UI function of nmRapp}
\usage{
nmRapp_ui(request)
}
\arguments{
\item{request}{default parameter necessary for specific UI as a function.}
}
\description{
Copyright (C) 2022 Battelle Memorial Institute

 This program is free software; you can redistribute it and/or modify
 it under the terms of the GNU General Public License as published by
 the Free Software Foundation; either version 2 of the License, or
 (at your option) any later version.

 This program is distributed in the hope that it will be useful,
 but WITHOUT ANY WARRANTY; without even the implied warranty of
 MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
 GNU General Public License for more details.

 You should have received a copy of the GNU General Public License along
 with this program; if not, write to the Free Software Foundation, Inc.,
 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
}
\details{
This main UI function references the UI components of modules specific to the processing and visualization
of experimental data and metadata. Refer to the documentation of each referenced module for details on its specific function.
A list of referenced modules is provided below:
1) xpmt_data_uploadUI()
2) xpmt_data_modifyUI()
3) xpmt_data_vizUI()
4) xpmt_metadata_vizUI()

This main UI function also creates the global UI format, i.e. navbarPage(), under which several different global tabs corresponding
to different stages of the NMR metabolite profiling workflow are made available. This environment is initialized with only
the experimental data upload and processing tab. The main server function, nmRapp_server(), dynamically updates the environment
with additional global tabs depending on the completion of the current tab requirements. For example, a "target metabolite editing"
tab is made available through nmRapp_server() only after the user has uploaded and processed their experimental data and metadata.
}
