% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opt_rDolphin.R
\name{opt_rDolphin}
\alias{opt_rDolphin}
\title{Optimization check for rDolphin}
\usage{
opt_rDolphin(imported_data)
}
\arguments{
\item{imported_data}{an rDolphin object, created using \code{\link{ppmData_to_rDolphin}}}
}
\description{
Optimization check for rDolphin
}
\author{
Allison Thompson
}
