% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rD_output_generator.R
\name{output_generator}
\alias{output_generator}
\title{rDolphin: output_generator}
\usage{
output_generator(
  signals_to_quantify,
  fitted_signals,
  Ydata,
  Xdata,
  signals_parameters,
  multiplicities,
  buck_step
)
}
\arguments{
\item{signals_to_quantify}{Indices of signals that were quantified}

\item{fitted_signals}{Intensity values of fitted signals}

\item{Ydata}{Dataframe of intensity values.}

\item{Xdata}{Dataframe of sample chemical shift (ppm) values.}

\item{signals_parameters}{vector of signal fitting parameters}

\item{multiplicities}{vector of signal multiplicity values}

\item{buck_step}{The difference between two adjacent chemical shift values (i.e. ppm resolution)}
}
\description{
rDolphin: output_generator
}
\details{
This function is based on code that was forked directly from rDolphin (https://github.com/danielcanueto/rDolphin/tree/master/R). This function formats
the output generated by the fitting routine and computes error metrics based on this output.
}
