% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.ppmData.R
\name{peak_alignment}
\alias{peak_alignment}
\title{Spectrum Alignment Pre-processing}
\usage{
peak_alignment(e_data)
}
\arguments{
\item{e_data}{a data.frame of expression data, first component of a ppmData object}
}
\value{
a data.frame containing aligned spectra with same format as `e_data`
}
\description{
This function takes the data.frame `e_data` and runs the CluPA spectrum alignment algorithm from the `speaq` R package
}
\author{
Natalie Winans
}
