% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rD_peakdet.R
\name{peakdet}
\alias{peakdet}
\title{rDolphin: peakdet}
\usage{
peakdet(v, delta, x = NULL)
}
\arguments{
\item{v}{a numeric vector where to search peaks.}

\item{delta}{numeric of length one; defining the local threshold for peak detection.}

\item{x}{a numeric vector the same length as v containing corresponding x-values for v.}
}
\value{
A list containing two data frames maxtab and mintab containing maxima and minima. Data frames contain two columns with indices in v (or corresponding
values in x if provided) and values in v.
}
\description{
rDolphin: peakdet
}
\details{
This function was forked directly from rDolphin (https://github.com/danielcanueto/rDolphin/tree/master/R). From rDolphin: 

A peak detection algorithm for R; finds local maxima and minima. Based on an algorithm for MATLAB by Eli Billauer (http://www.billauer.co.il/peakdet.html).
This function is released to the public domain; any use is allowed.
}
