% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rD_peakpvoigt.R
\name{peakpvoigt}
\alias{peakpvoigt}
\title{rDolphin: peakpvoigt}
\usage{
peakpvoigt(x, p)
}
\arguments{
\item{x}{}

\item{p}{}
}
\description{
rDolphin: peakpvoigt
}
\details{
This function was forked directly from rDolphin (https://github.com/danielcanueto/rDolphin/tree/master/R). This is the
pseudovoigt lineshape function.
}
