% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_plot_xpmt_data.R
\name{plot_xpmt_data}
\alias{plot_xpmt_data}
\title{Plot uploaded experimental data}
\usage{
plot_xpmt_data(xpmt_data, sourceid, sample_to_plot = NULL, brushed_data = NULL)
}
\arguments{
\item{xpmt_data}{A dataframe with at least two columns. See details.}

\item{sourceid}{A string specifying the id of the generated plotly figure.}

\item{sample_to_plot}{A string specifying denoting which sample should be plotted (for the main plot).}

\item{brushed_data}{plotly eventData corresponding to a brush event. This is used to inform subplot generation.}
}
\value{
A plotly object
}
\description{
Copyright (C) 2022 Battelle Memorial Institute

 This program is free software; you can redistribute it and/or modify
 it under the terms of the GNU General Public License as published by
 the Free Software Foundation; either version 2 of the License, or
 (at your option) any later version.

 This program is distributed in the hope that it will be useful,
 but WITHOUT ANY WARRANTY; without even the implied warranty of
 MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
 GNU General Public License for more details.

 You should have received a copy of the GNU General Public License along
 with this program; if not, write to the Free Software Foundation, Inc.,
 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
}
\details{
This function generates a plotly figure of the user-uploaded experimental data. The first column of the dataframe
required by the argument "xpmt_data" should contain the experimental PPM values. Subsequent columns should contain the
corresponding intensity values for each experimental sample.
}
