% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppmData_to_rDolphin.R
\name{ppmData_to_rDolphin}
\alias{ppmData_to_rDolphin}
\title{Convert Data to rDolphin input type}
\usage{
ppmData_to_rDolphin(
  ppmData,
  metabs,
  normalization = 0,
  alignment = 0,
  nmr_folder_path = "",
  data_index_1d = "",
  proc_no = "",
  default_suppressions = "",
  bucket_resolution = "",
  specific_params = ""
)
}
\arguments{
\item{ppmData}{a \code{ppmData} object}

\item{metabs}{metabolite information, from \code{get_reference_data} for example}

\item{normalization}{Numeric indicating which normalization to perform - 0 = No normalization; 1 = Eretic; 2 = TSP; 3 = Creatinine; 4 = Spectra Sum; 5 = PQN. Default is 0.}

\item{alignment}{Numeric indicating which alignment to perform - 0 = No alignment; 1 = Glucose; 2 = TSP; 3 = Formate. Default is 0.}

\item{nmr_folder_path}{Folder where all sample NMR folders are stored, if spectra are generated from Bruker files; default is "", meaning nothing is imported.}

\item{data_index_1d}{Name of 1D data folder inside each sample NMR folder (usually along the lines of 11, 12, 13...), if spectra are generated from Bruker files. Default is "", meaning nothing is imported.}

\item{proc_no}{Name of the proc_no folder inside each 1D data folder (usually this is 1), if spectra are generated from Bruker files. Default is "", meaning nothing is imported.}

\item{default_suppressions}{Spectrum regions suppressed, with the format 9.5-6.1;5.1-4.5. Default is "", no suppressions.}

\item{bucket_resolution}{PPM for bucket resolutions; default is ""}

\item{specific_params}{Path of file to change rDolphin internal parameters if desired}
}
\value{
Returns an rDolphin data object
}
\description{
Copyright (C) 2022 Battelle Memorial Institute

 This program is free software; you can redistribute it and/or modify
 it under the terms of the GNU General Public License as published by
 the Free Software Foundation; either version 2 of the License, or
 (at your option) any later version.

 This program is distributed in the hope that it will be useful,
 but WITHOUT ANY WARRANTY; without even the implied warranty of
 MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
 GNU General Public License for more details.

 You should have received a copy of the GNU General Public License along
 with this program; if not, write to the Free Software Foundation, Inc.,
 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.

Converts a 'ppmData' object into input files for rDolphin
}
\details{
Objects of class 'ppmData' contain some attributes that are referenced by downstream functions, namely rDolphin. These attributes can be changed from their default value by manual specification. A list of these attributes as well as their default values are as follows:
\tabular{ll}{
nmr_folder_path \tab Folder where all sample NMR folders are stored, if spectra are generated from Bruker files; default is "", meaning nothing is imported. \cr
\tab \cr
data_index_1d \tab Name of 1D data folder inside each sample NMR folder (usually along the lines of 11, 12, 13...), if spectra are generated from Bruker files. Default is "", meaning nothing is imported. \cr
\tab \cr
proc_no \tab Name of the proc_no folder inside each 1D data folder (usually this is 1), if spectra are generated from Bruker files. Default is "", meaning nothing is imported. \cr
\tab \cr
normalization \tab Numeric indicating which normalization to perform - 0 = No normalization; 1 = Eretic; 2 = TSP; 3 = Creatinine; 4 = Spectra Sum; 5 = PQN. Default is 0. \cr
\tab \cr
alignment \tab Numeric indicating which alignment to perform - 0 = No alignment; 1 = Glucose; 2 = TSP; 3 = Formate. Default is 0. \cr
\tab \cr
default_suppressions \tab Spectrum regions suppressed, with the format 9.5-6.1;5.1-4.5. Default is "", no suppressions. \cr
\tab \cr
bucket_resolution \tab PPM for bucket resolutions; default is "". \cr
\tab \cr
specific_params \tab Path of file to change rDolphin internal parameters if desired. \cr
\tab \cr
}
}
\author{
Allison Thompson
}
