% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_module_profiling.R
\name{profiling_completeviewUI}
\alias{profiling_completeviewUI}
\title{Module: UI element to display trelliscope plot of fitted spectra}
\usage{
profiling_completeviewUI(id)
}
\arguments{
\item{id}{A string denoting the namespace id.}
}
\description{
Copyright (C) 2022 Battelle Memorial Institute

 This program is free software; you can redistribute it and/or modify
 it under the terms of the GNU General Public License as published by
 the Free Software Foundation; either version 2 of the License, or
 (at your option) any later version.

 This program is distributed in the hope that it will be useful,
 but WITHOUT ANY WARRANTY; without even the implied warranty of
 MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
 GNU General Public License for more details.

 You should have received a copy of the GNU General Public License along
 with this program; if not, write to the Free Software Foundation, Inc.,
 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
}
\details{
This is one of the UI components for the module created to handle all functions related to profiling
of experimental data, based on user edits to the fitting parameters of a pre-specified list of
reference (target) metabolites. The value provided for 'id' should be identical across the following:
profiling_completeviewUI(), profiling_detailedviewUI(), and profilingServer().

This module component provides the UI element that allow users to:
1) View (within a trelliscope display) the fitted spectra over the sample spectra for all samples and target metabolites
}
