% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_module_profiling.R
\name{profiling_detailedviewUI}
\alias{profiling_detailedviewUI}
\title{Module: UI element to display trelliscope plot of fitted spectra}
\usage{
profiling_detailedviewUI(id)
}
\arguments{
\item{id}{A string denoting the namespace id.}
}
\description{
Module: UI element to display trelliscope plot of fitted spectra
}
\details{
This is one of the UI components for the module created to handle all functions related to profiling
of experimental data, based on user edits to the fitting parameters of a pre-specified list of
reference (target) metabolites. The value provided for 'id' should be identical across the following:
profiling_completeviewUI(), profiling_detailedviewUI(), and profilingServer().

This module component provides the UI elements that allow users to:
1) Select from several options to control the shown display
2) View interactively the fitted spectra of a given metabolite over a given sample spectrum
}
