% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_module_profiling.R
\name{profiling_prequantUI}
\alias{profiling_prequantUI}
\title{Module: UI element displaying aggregated ROIs based on final edited metabolite fitting data}
\usage{
profiling_prequantUI(id)
}
\arguments{
\item{id}{A string denoting the namespace id.}
}
\description{
Copyright (C) 2022 Battelle Memorial Institute

 This program is free software; you can redistribute it and/or modify
 it under the terms of the GNU General Public License as published by
 the Free Software Foundation; either version 2 of the License, or
 (at your option) any later version.

 This program is distributed in the hope that it will be useful,
 but WITHOUT ANY WARRANTY; without even the implied warranty of
 MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
 GNU General Public License for more details.

 You should have received a copy of the GNU General Public License along
 with this program; if not, write to the Free Software Foundation, Inc.,
 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
}
\details{
This is one of the UI components for the module created to handle all profiling-related functions.
The value provided for 'id' should be identical across the following: profiling_quant_sidebarUI(), profiling_prequantUI(),
profiling_completeviewUI(), profiling_detailedviewUI(), and profilingServer().

This module component displays the set of target metabolite data that will be used for profiling, aggregated into ROIs.
These data reflect any and all edits made to the target metabolite data and are displayed within a searchable, filterable table.
}
