% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_ProxyUpdate_refmet_tabplot.R
\name{ProxyUpdate_refmet_tabplot}
\alias{ProxyUpdate_refmet_tabplot}
\title{Simultaneously update reference metabolite datatable and plot via proxies}
\usage{
ProxyUpdate_refmet_tabplot(tabproxy, pltproxy, newdat)
}
\arguments{
\item{tabproxy}{A datatable proxy object. See details.}

\item{pltproxy}{A plotly proxy object. See details.}

\item{newdat}{A dataframe. See details.}
}
\description{
Copyright (C) 2022 Battelle Memorial Institute

 This program is free software; you can redistribute it and/or modify
 it under the terms of the GNU General Public License as published by
 the Free Software Foundation; either version 2 of the License, or
 (at your option) any later version.

 This program is distributed in the hope that it will be useful,
 but WITHOUT ANY WARRANTY; without even the implied warranty of
 MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
 GNU General Public License for more details.

 You should have received a copy of the GNU General Public License along
 with this program; if not, write to the Free Software Foundation, Inc.,
 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
}
\details{
This function is responsible for updating the reference metabolite datatable and accompanying plot
according to the changes made interactively to each by the user. The datatable proxy object required by
tabproxy corresponds to the datatable displaying the reference metabolite fitting information. See
ref_data_ROIeditingUI() and ref_data_editingServer() for additional details.
The plotly proxy object corresponds to the plotly figured generated by plot_refmet_data(). The dataframe required by
newdat should contain all of the fitting parameter information specific to the peak centers of a selected
reference metabolite. This dataframe should have the following named columns, in the following order: "ROI left edge (ppm)",
"ROI right edge (ppm)", "Quantification Mode", "Metabolite", "Quantification Signal", "Chemical shift(ppm)",
"Chemical shift tolerance (ppm)", "Half bandwidth (Hz)", "Multiplicity", "J coupling (Hz)", "Roof effect",
"Quantify", "HMDB_code", and "rowid".
}
