% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_module_ref_data_modify.R
\name{ref_data_add_delUI}
\alias{ref_data_add_delUI}
\title{Module: UI elements specific to adding/removing target metabolites}
\usage{
ref_data_add_delUI(id)
}
\arguments{
\item{id}{A string denoting the namespace id.}
}
\description{
Module: UI elements specific to adding/removing target metabolites
}
\details{
This is one of the UI components for the module created to handle all editing of reference (target) metabolite data.
The value provided for 'id' should be identical across the following: ref_data_ROIeditingUI(), ref_data_add_delUI(),
ref_data_quantTab(), ref_data_profileUI(), ref_data_global_fittingparamsUI(), and ref_data_editingServer().

This module component provides the UI elements that allow users to:
1) Specify metabolite(s) to add or remove
2) Add/remove specified metabolite(s)

This module component also displays our internal reference database within a searchable, filterable table.
}
