% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_module_ref_data_modify.R
\name{ref_data_global_fittingparamsUI}
\alias{ref_data_global_fittingparamsUI}
\title{Module: UI element displaying global profiling options}
\usage{
ref_data_global_fittingparamsUI(id)
}
\arguments{
\item{id}{A string denoting the namespace id.}
}
\description{
Module: UI element displaying global profiling options
}
\details{
This is one of the UI components for the module created to handle all editing of reference (target) metabolite data.
The value provided for 'id' should be identical across the following: ref_data_ROIeditingUI(), ref_data_add_delUI(),
ref_data_quantTab(), ref_data_profileUI(), ref_data_global_fittingparamsUI(), and ref_data_editingServer().

This module component displays all global profiling options that are listed under rDolphin's "program_parameters."
}
