% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_ref_data_modify.R
\name{ref_data_profileUI}
\alias{ref_data_profileUI}
\title{Module: Profiling button}
\usage{
ref_data_profileUI(id)
}
\arguments{
\item{id}{A string denoting the namespace id.}
}
\description{
Module: Profiling button
}
\details{
This is one of the UI components for the module created to handle all editing of reference (target) metabolite data.
The value provided for 'id' should be identical across the following: ref_data_ROIeditingUI(), ref_data_add_delUI(),
ref_data_quantTab(), ref_data_profileUI(), and ref_data_editingServer().

This module component provides the UI elements that allow users to initiate metabolite profiling
}
