% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_module_ref_data_upload.R
\name{ref_data_uploadUI}
\alias{ref_data_uploadUI}
\title{Module: UI elements specific to reference metabolite data uploading}
\usage{
ref_data_uploadUI(id, ref_db)
}
\arguments{
\item{id}{A string denoting the namespace id.}

\item{ref_db}{Dataframe of internal database of reference metabolite data.}
}
\description{
Copyright (C) 2022 Battelle Memorial Institute

 This program is free software; you can redistribute it and/or modify
 it under the terms of the GNU General Public License as published by
 the Free Software Foundation; either version 2 of the License, or
 (at your option) any later version.

 This program is distributed in the hope that it will be useful,
 but WITHOUT ANY WARRANTY; without even the implied warranty of
 MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
 GNU General Public License for more details.

 You should have received a copy of the GNU General Public License along
 with this program; if not, write to the Free Software Foundation, Inc.,
 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
}
\details{
This is the UI component for the module created to handle all uploading/specification of reference (target) metabolite data.
The value provided for 'id' should be identical across ref_data_uploadUI() and ref_data_uploadServer().

This module component provides the UI elements that allow users to:
1) Select an import method for the target metabolite data. Users may choose to either upload a .xlsx file
containing a column of CAS numbers or users may choose to manually specify metabolites by name. Note that
autocompletion functionality is available for manual specification.
2) Import target (reference) metabolite data via the selected import method.
}
