% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refmet_data_change_fromtab.R
\name{refmet_data_change_fromtab}
\alias{refmet_data_change_fromtab}
\title{Record any changes made to target reference metabolite data based on plot edits}
\usage{
refmet_data_change_fromtab(
  dspedt_refmet_data,
  changed_row,
  change,
  edtd_colname,
  round_num = 3
)
}
\arguments{
\item{dspedt_refmet_data}{A dataframe containing fitting parameter information specific to a single target metabolite. See details.}

\item{changed_row}{An integer denoting the row of the table edit.}

\item{change}{The updated value. May be either a string or numeric depending on the table column edited.}

\item{edtd_colname}{A string denoting the name of the column of the table edit}

\item{round_num}{An integer specifying the number of decimals to round changed (numeric) values to.}
}
\value{
A dataframe containing fitting parameter information specific to a single target metabolite, updated to reflect the
table-induced changes
}
\description{
Record any changes made to target reference metabolite data based on plot edits
}
\details{
The dspedt_refmet_data dataframe should have the following named columns, in the following order: "ROI left edge (ppm)", 
"ROI right edge (ppm)", "Quantification Mode", "Metabolite", "Quantification Signal", "Chemical shift(ppm)", 
"Chemical shift tolerance (ppm)", "Half bandwidth (Hz)", "Multiplicity", "J coupling (Hz)", "Roof effect", 
"Quantify", "HMDB_code", and "rowid". Only the peak centers of a single, user-selected metabolite should be represented
within this dataframe.
}
