% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{refmets}
\alias{refmets}
\title{Reference Metabolites}
\format{
A data frame with 2,224 rows and 6 columns:
\describe{
  \item{Entry_ID}{}
  \item{CASno}{}
  \item{Field_strength}{}
  \item{Solute}{}
  \item{Solvent}{}
  \item{pH}{}
}
}
\usage{
refmets
}
\description{
Copyright (C) 2022 Battelle Memorial Institute

 This program is free software; you can redistribute it and/or modify
 it under the terms of the GNU General Public License as published by
 the Free Software Foundation; either version 2 of the License, or
 (at your option) any later version.

 This program is distributed in the hope that it will be useful,
 but WITHOUT ANY WARRANTY; without even the implied warranty of
 MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
 GNU General Public License for more details.

 You should have received a copy of the GNU General Public License along
 with this program; if not, write to the Free Software Foundation, Inc.,
 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.

A dataset containing reference metabolite data
}
\keyword{datasets}
