% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_filter_ppm.R
\name{remove_filter_ppm}
\alias{remove_filter_ppm}
\title{Remove PPM Filter}
\usage{
remove_filter_ppm(ppmData, filters)
}
\arguments{
\item{ppmData}{a `ppmData` object, consisting of a list containing the data tables `e_data` and `f_data`, upon which a PPM filter has previously been applied using the function \code{\link{filter_ppm}}}

\item{filters}{integer vector specifying which PPM filter(s) to remove. Can be accessed using \code{\link{attr}} function.}
}
\value{
a `ppmData` object with previously removed rows restored to `e_data` and PPM filter attributes set to `NULL`
}
\description{
Removes a previously applied PPM filter by restoring removed rows and clearing 'filters' attribute
}
\examples{
\dontrun{
remove_filter_ppm(mydata, filters = c(1,2))
}

}
\author{
Natalie Winans
}
