% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_reference_data.R
\name{roi_ref_export}
\alias{roi_ref_export}
\title{Wrap the generating and exporting functions into one}
\usage{
roi_ref_export(
  name_list = NULL,
  cas_list = NULL,
  return_metabs = "exact_match",
  solvent_type = NULL,
  ph = NULL,
  half_bandwidth = 1.4,
  roi_tol = 0.02,
  instrument_strength = NULL
)
}
\arguments{
\item{name_list}{list of metabolite names if cas numbers are not provided}

\item{cas_list}{list of CAS registry numbers written as characters if metabolite names are not provided}

\item{return_metabs}{string, must be one of "exact_match", "nearest_match", or "all". Defaults to "exact_match". If "exact_match", returns metabolites that exactly match specified experimental conditions (solvent type, pH, and instrument strength). If "nearest_match", returns nearest match metabolites, sorted by Euclidean distance. If "all", returns all entries corresponding to supplied CAS numbers, experimental conditions ignored.}

\item{solvent_type}{the experimental solvent, choose from available solvents 'D2O', 'H2O', ...}

\item{ph}{the experimental pH}

\item{half_bandwidth}{This will be set to 1.4 unless otherwise specified}

\item{roi_tol}{the tolerance to encompass the region of interest value}

\item{instrument_strength}{the field strength of the instrument used to collect the data}
}
\value{
a data frame formatted for rDolphin use to be exported as excel sheet
}
\description{
Wrap the generating and exporting functions into one
}
