% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_rDolphin.R
\name{run_rDolphin}
\alias{run_rDolphin}
\title{Wrapper for running rDolphin}
\usage{
run_rDolphin(
  imported_data,
  dir = NULL,
  optimization = TRUE,
  spectra_to_profile = NULL
)
}
\arguments{
\item{imported_data}{an rDolphin data object, often created by \code{ppmData_to_rDolphin}}

\item{dir}{Directory where to save results, default is NA, which will create a folder in your working directory.}

\item{optimization}{Logical, should profiling quality be maximized through analysis of signal parameters. Default is TRUE.}

\item{spectra_to_profile}{A vector of spectrum indices. Defaults to NULL, profiling all spectra.}
}
\description{
Runs rDolphin using imported_data (the output from \code{\link{ppmData_to_rDolphin}}) and \code{\link{opt_rDolphin}}
}
\author{
Allison Thompson
}
