% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rD_save_output.R
\name{save_output}
\alias{save_output}
\title{rDolphin: save_output}
\usage{
save_output(
  spectrum_index,
  signals_codes,
  results_to_save,
  buck_step,
  final_output
)
}
\arguments{
\item{spectrum_index}{Index of sample spectrum to be fitted}

\item{signals_codes}{Signal indices to be fitted}

\item{results_to_save}{Final results list}

\item{buck_step}{The difference between two adjacent chemical shift values (i.e. ppm resolution)}

\item{final_output}{Dataframe for final quantifications}
}
\description{
rDolphin: save_output
}
\details{
This function is based on code that was forked directly from rDolphin (https://github.com/danielcanueto/rDolphin/tree/master/R). This function saves
the formatted output of the quantification generated by the automatic profiling routine.
}
