% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_reference_data.R
\name{show_file}
\alias{show_file}
\title{Imports whole STAR file for a given entry id from a locally maintained respository of JSON formated BMRM-STAR files}
\usage{
show_file(ID)
}
\arguments{
\item{ID}{single BMRB ID; For metabolomics entries it is just a number with bmrb prefix (example: 'bmse000028');}
}
\value{
R list that contains STAR formatted data for a given entry ID
}
\description{
Copyright (C) 2022 Battelle Memorial Institute

 This program is free software; you can redistribute it and/or modify
 it under the terms of the GNU General Public License as published by
 the Free Software Foundation; either version 2 of the License, or
 (at your option) any later version.

 This program is distributed in the hope that it will be useful,
 but WITHOUT ANY WARRANTY; without even the implied warranty of
 MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
 GNU General Public License for more details.

 You should have received a copy of the GNU General Public License along
 with this program; if not, write to the Free Software Foundation, Inc.,
 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
}
\examples{
# file <- show_file("path/to/json/","bmse000028")
# Downloads entire STAR file from BMRB metabolomics database
}
\seealso{
\code{\link{get_spectra_data}}
}
