% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subplot_refmet_data.R
\name{subplot_refmet_data}
\alias{subplot_refmet_data}
\title{Plot selected target (reference) metabolite data ROI's over brushed region of experimental data}
\usage{
subplot_refmet_data(
  xpmt_data,
  brushed_data,
  sample_to_plot,
  ROI_lines,
  ROI_annots,
  sourceid
)
}
\arguments{
\item{xpmt_data}{A dataframe with at least two columns. See details.}

\item{brushed_data}{plotly eventData corresponding to a brush event.}

\item{sample_to_plot}{A string specifying denoting which sample is featured in the main plot.}

\item{ROI_lines}{A list containing information necessary to create plotly shapes. See details.}

\item{ROI_annots}{A list containing information necessary to create plotly annotations. See details.}

\item{sourceid}{A string specifying the id of the generated plotly figure.}
}
\value{
A plotly object
}
\description{
Plot selected target (reference) metabolite data ROI's over brushed region of experimental data
}
\details{
This function generates a plotly figure specific to the experimental data at the brushed (selected) region of the main
plot. The first column of the dataframe required by the argument "xpmt_data" should contain the experimental PPM values. 
Subsequent columns should contain the corresponding intensity values for each experimental sample. 
The "ROI_lines" and "ROI_annots" lists should have as many elements as there are peak centers for the given target metabolite. 
Each list element of "ROI_lines" and "ROI_annots" are lists containing several parameters that describe how interactive lines
and annotations should be drawn over the generated plotly figure. See ROI_line_gen() and ROI_annot_gen() for further details.
}
