% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trel_plot.R
\name{trel_plot}
\alias{trel_plot}
\title{Create Trelliscope plot}
\usage{
trel_plot(plot.data)
}
\arguments{
\item{plot.data}{Resultant object from \code{\link{format_plotting}}}
}
\value{
Plot object for Trelliscope Plotting
}
\description{
Create Trelliscope plot
}
\author{
Allison Thompson
}
