% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_xpmt_data_upload.R
\name{xpmt_data_uploadServer}
\alias{xpmt_data_uploadServer}
\title{Module: Server functions specific to experimental data uploading}
\usage{
xpmt_data_uploadServer(id)
}
\arguments{
\item{id}{A string denoting the namespace id.}
}
\value{
A reactive object containing the uploaded experimental data and parameters saved as an nmRanalysis ppmData object.
}
\description{
Module: Server functions specific to experimental data uploading
}
\details{
This is the UI component for the module created to handle the uploading of experimental data.
The value provided for 'id' should be identical across xpmt_data_uploadUI() and xpmt_data_uploadServer().

This module component provides the back-end code that:
1) Processes uploaded data based on the provided experimental conditions. Data are saved as an nmRanalysis ppmData object.
}
