% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_module_xpmt_data_upload.R
\name{xpmt_data_uploadUI}
\alias{xpmt_data_uploadUI}
\title{Module: UI elements specific to experimental data uploading}
\usage{
xpmt_data_uploadUI(id, ref_db)
}
\arguments{
\item{id}{A string denoting the namespace id.}

\item{ref_db}{Dataframe of internal database of reference metabolite data.}
}
\description{
Copyright (C) 2022 Battelle Memorial Institute

 This program is free software; you can redistribute it and/or modify
 it under the terms of the GNU General Public License as published by
 the Free Software Foundation; either version 2 of the License, or
 (at your option) any later version.

 This program is distributed in the hope that it will be useful,
 but WITHOUT ANY WARRANTY; without even the implied warranty of
 MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
 GNU General Public License for more details.

 You should have received a copy of the GNU General Public License along
 with this program; if not, write to the Free Software Foundation, Inc.,
 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
}
\details{
This is the UI component for the module created to handle the uploading of experimental data.
The value provided for 'id' should be identical across xpmt_data_uploadUI() and xpmt_data_uploadServer().

This module component provides the UI elements that allow users to:
1) Upload the .csv file containing the experimental PPM and Intensity data for all samples. The first column of
this datafile should provide the PPM values and subsequent columns the corresponding intensity values for each
sample.
2) Upload the .csv file containing experimental metadata. Currently, the only metadata processed are sample labels.
3) Specify experimental pH, instrument strength, and the solvent used. The available choices for each of these
"experimental conditions" is constrained by the conditions available in our internal database.
}
