% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_xpmt_data_viz.R
\name{xpmt_data_vizoptionsUI}
\alias{xpmt_data_vizoptionsUI}
\title{Module: UI elements specific to data visualization options (e.g. filters, subplots)}
\usage{
xpmt_data_vizoptionsUI(id)
}
\arguments{
\item{id}{A string denoting the namespace id.}
}
\description{
Module: UI elements specific to data visualization options (e.g. filters, subplots)
}
\details{
This is the UI component for the module created to handle editing of uploaded experimental data.
The value provided for 'id' should be identical across xpmt_data_vizUI(), xpmt_data_vizoptionsUI(), 
xpmt_metadata_vizUI, and xpmt_data_vizServer().

This module component provides the dynamic UI elements that allow users to:
1) Select which sample spectrum should be displayed on the main plotly plot
2) Toggle the option for subplot display of any selected plot region. This subplot
   will shown the intensities across all sample spectra at the selected region.
3) Apply or remove uploaded data filter(s)

Note that since this UI component is dynamic, its code is found within the corresponding server component of the 
module, xpmt_data_vizServer().
}
