% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_module_xpmt_data_viz.R
\name{xpmt_data_vizServer}
\alias{xpmt_data_vizServer}
\title{Module: Server functions specific to experimental data and metadata visualization}
\usage{
xpmt_data_vizServer(id, xpmt_data)
}
\arguments{
\item{id}{A string denoting the namespace id.}

\item{xpmt_data}{The reactive object returned by the xpmt_data_uploadServer() server module containing
the experimental data as a nmRanalysis ppmData object.}
}
\value{
A reactive object containing the uploaded experimental data with any modifications (i.e. filters) applied. Note that
any applied filters are stored as an attribute of the object. The attribute contains both the filtered range and the set of data
that was filtered within this range.
}
\description{
Copyright (C) 2022 Battelle Memorial Institute

 This program is free software; you can redistribute it and/or modify
 it under the terms of the GNU General Public License as published by
 the Free Software Foundation; either version 2 of the License, or
 (at your option) any later version.

 This program is distributed in the hope that it will be useful,
 but WITHOUT ANY WARRANTY; without even the implied warranty of
 MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
 GNU General Public License for more details.

 You should have received a copy of the GNU General Public License along
 with this program; if not, write to the Free Software Foundation, Inc.,
 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
}
\details{
This is the UI component for the module created to handle the visualization of experimental data and metadata.
The value provided for 'id' should be identical across xpmt_data_vizUI(), xpmt_data_vizoptionsUI(),
and xpmt_data_vizServer().

This module component provides the back-end code that:
1) Generates the plotly figure of experimental data
2) Generates the datatable of experimental metadata
3) Creates and displays subplots corresponding to selected regions
4) Applies or removes filters to the data
}
