% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_module_xpmt_data_viz.R
\name{xpmt_data_vizUI}
\alias{xpmt_data_vizUI}
\title{Module: UI elements specific to experimental data visualization}
\usage{
xpmt_data_vizUI(id)
}
\arguments{
\item{id}{A string denoting the namespace id.}
}
\description{
Module: UI elements specific to experimental data visualization
}
\details{
This is the UI component for the module created to handle the visualization of experimental data.
The value provided for 'id' should be identical across xpmt_data_vizUI(), xpmt_data_vizoptionsUI(),
xpmt_metadata_vizUI, and xpmt_data_vizServer().

This module component provides the UI elements that allow users to:
1) Visualize uploaded experimental data through an interactive plotly figure
2) Visualize uploaded experimental data through a searchable datatable.
}
