% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_reference_data.R
\name{as.bmseListFromName}
\alias{as.bmseListFromName}
\title{Verifying an entry exists for a given metabolite name}
\usage{
as.bmseListFromName(
  name_list,
  return_metabs = "all",
  solvent_type = NULL,
  ph = NULL,
  instrument_strength,
  temperature = NULL,
  concentration = NULL
)
}
\arguments{
\item{name_list}{list of metabolite names}

\item{return_metabs}{string, must be one of "exact_match" or "all". Defaults to "all". If "exact_match", returns metabolites that exactly match specified experimental conditions (solvent type, pH, and instrument strength). If "all", returns all entries corresponding to supplied CAS numbers, experimental conditions ignored.}

\item{solvent_type}{choose from available solvents 'D2O', 'H2O', ...}

\item{ph}{numerical value specifying pH of the experimental conditions}

\item{instrument_strength}{numeric value specifying experimental instrument strength}

\item{temperature}{numeric value specifying the temperature}

\item{concentration}{numeric value specifying the concentration}
}
\value{
a list of BMSE database entry numbers
}
\description{
Copyright (C) 2022 Battelle Memorial Institute

 This program is free software; you can redistribute it and/or modify
 it under the terms of the GNU General Public License as published by
 the Free Software Foundation; either version 2 of the License, or
 (at your option) any later version.

 This program is distributed in the hope that it will be useful,
 but WITHOUT ANY WARRANTY; without even the implied warranty of
 MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
 GNU General Public License for more details.

 You should have received a copy of the GNU General Public License along
 with this program; if not, write to the Free Software Foundation, Inc.,
 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
}
\examples{
# test <- as.bmseListFromName(list("ATP","Maltose" "Oxalate"), solvent_type = 'D2O', ph = 7.4)

}
