% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bmse_associations}
\alias{bmse_associations}
\title{BMSE Associations}
\format{
A data frame with 2224 rows and 9 columns:
\describe{
  \item{Entry_ID}{BMSE identifier}
  \item{CASno}{CAS number}
  \item{Field_strength}{}
  \item{Solute}{}
  \item{Solvent}{Experimental conditions}
  \item{Reference}{}
  \item{pH}{Experimental conditions}
  \item{Temperature}{Experimental conditions}
  \item{Concentration}{Experimental Conditions}
}
}
\usage{
bmse_associations
}
\description{
Copyright (C) 2022 Battelle Memorial Institute

 This program is free software; you can redistribute it and/or modify
 it under the terms of the GNU General Public License as published by
 the Free Software Foundation; either version 2 of the License, or
 (at your option) any later version.

 This program is distributed in the hope that it will be useful,
 but WITHOUT ANY WARRANTY; without even the implied warranty of
 MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
 GNU General Public License for more details.

 You should have received a copy of the GNU General Public License along
 with this program; if not, write to the Free Software Foundation, Inc.,
 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.

A dataset containing metabolite data
}
\keyword{datasets}
