% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{e_data}
\alias{e_data}
\title{Example e_data}
\format{
A data frame with 262144 rows and 5 columns:
\describe{
  \item{PPM}{Observed PPMs in NMR spectra}
  \item{AA_1x_a_20201209.17.fid.1}{Sample data at each PPM}
  \item{AA_1x_a_20201209.18.fid.2}{Sample data at each PPM}
  \item{AA_1x_a_20201209.19.fid.3}{Sample data at each PPM}
  \item{AA_1x_a_20201209.20.fid.4}{Sample data at each PPM}
}
}
\usage{
e_data
}
\description{
A dataset containing NMR spectral data over 4 samples
}
\keyword{datasets}
