% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edit_reference_data.R
\name{edit_reference_data}
\alias{edit_reference_data}
\title{Edit Reference Data}
\usage{
edit_reference_data(
  roi_df,
  metabolite_name,
  ppm,
  tolerance = 0.05,
  parameter_name,
  parameter_value
)
}
\arguments{
\item{roi_df}{data frame returned from roi_ref_export() that contains the rDolphin reference data}

\item{metabolite_name}{character vector with one or more entries}

\item{ppm}{numeric vector with one or more entries; chemical shift of the parameter to be changed}

\item{tolerance}{chemical shift tolerance to search for a peak within; default is 0.05}

\item{parameter_name}{character vector with one or more entries; parameters to be edited}

\item{parameter_value}{character vector with one or more entries; the new value a parameter will be set to with the specified identifiers}
}
\value{
an edited data.frame formatted for rDolphin use
}
\description{
Copyright (C) 2022 Battelle Memorial Institute

 This program is free software; you can redistribute it and/or modify
 it under the terms of the GNU General Public License as published by
 the Free Software Foundation; either version 2 of the License, or
 (at your option) any later version.

 This program is distributed in the hope that it will be useful,
 but WITHOUT ANY WARRANTY; without even the implied warranty of
 MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
 GNU General Public License for more details.

 You should have received a copy of the GNU General Public License along
 with this program; if not, write to the Free Software Foundation, Inc.,
 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
}
