% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_metabs}
\alias{example_metabs}
\title{Example metabolites}
\format{
A data frame with 10 rows and 12 columns:
\describe{
  \item{ROI left edge (ppm)}{Left boundary of region of interest}
  \item{ROI right edge (ppm)}{Right boundary of region of interest}
  \item{Quantification Mode}{}
  \item{Metabolite}{}
  \item{Quantification Signal}{}
  \item{Chemical shift(ppm)}{}
  \item{Chemical shift tolerance (ppm)}{}
  \item{Half bandwidth (Hz)}{}
  \item{Multiplicity}{}
  \item{J coupling (Hz)}{}
  \item{Roof effect}{}
  \item{HMDB_code}{}
}
}
\usage{
example_metabs
}
\description{
A dataset containing example metabolite data
}
\keyword{datasets}
