% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_reference_data.R
\name{export_roi_file}
\alias{export_roi_file}
\title{Format spectra data into a proper rDolphin ROI reference file}
\usage{
export_roi_file(
  spectra_df,
  return_metabs = "exact_match",
  half_bandwidth = 1.4,
  roi_tol = 0.02,
  instrument_strength
)
}
\arguments{
\item{spectra_df}{data frame of the spectra from get_spectra_data() function}

\item{return_metabs}{string, must be one of "exact_match", "nearest_match", or "all". Defaults to "exact_match". If "exact_match", returns metabolites that exactly match specified experimental conditions (solvent type, pH, and instrument strength). If "nearest_match", returns nearest match metabolites, sorted by Euclidean distance. If "all", returns all entries corresponding to supplied CAS numbers, experimental conditions ignored.}

\item{half_bandwidth}{This will be set to 1.4 unless otherwise specified}

\item{roi_tol}{The ROI tolerance, length from the center of the peak to each edge. Default is 0.02.}

\item{instrument_strength}{the field strength of the instrument used to collect the data}
}
\value{
a data frame formatted for r dolphin use to be exported as excel sheet
}
\description{
Copyright (C) 2022 Battelle Memorial Institute

 This program is free software; you can redistribute it and/or modify
 it under the terms of the GNU General Public License as published by
 the Free Software Foundation; either version 2 of the License, or
 (at your option) any later version.

 This program is distributed in the hope that it will be useful,
 but WITHOUT ANY WARRANTY; without even the implied warranty of
 MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
 GNU General Public License for more details.

 You should have received a copy of the GNU General Public License along
 with this program; if not, write to the Free Software Foundation, Inc.,
 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
}
