% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{f_data}
\alias{f_data}
\title{Example f_data}
\format{
A data frame with 4 rows and 2 columns:
\describe{
  \item{Sample}{Unique sample identifiers}
  \item{Experiment}{Experiment number}
}
}
\usage{
f_data
}
\description{
A dataset containing sample metadata
}
\keyword{datasets}
