% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_ppm.R
\name{filter_ppm}
\alias{filter_ppm}
\title{PPM Filtering}
\usage{
filter_ppm(ppmData, range)
}
\arguments{
\item{ppmData}{a list object from containing the data tables `e_data` and `f_data`, created using the function \code{\link{as.ppmData}}}

\item{range}{a list specifying the minimum and maximum value(s) in the range(s) to be filtered. Multiple ranges may be specified. Ranges are inclusive.}
}
\value{
`ppmData` object with 'filters' attribute including range and `data.frame` with removed rows
}
\description{
Copyright (C) 2022 Battelle Memorial Institute

 This program is free software; you can redistribute it and/or modify
 it under the terms of the GNU General Public License as published by
 the Free Software Foundation; either version 2 of the License, or
 (at your option) any later version.

 This program is distributed in the hope that it will be useful,
 but WITHOUT ANY WARRANTY; without even the implied warranty of
 MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
 GNU General Public License for more details.

 You should have received a copy of the GNU General Public License along
 with this program; if not, write to the Free Software Foundation, Inc.,
 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.

Filters out specified PPM ranges from the dataset
}
\examples{
\dontrun{
filter_ppm(mydata, range = list(min = 2, max = 3))
filter_ppm(mydata, range = list(min = c(1,3), max = c(2,4)))
}

}
\author{
Allison Thompson, Natalie Winans
}
