% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rD_fitting_prep_2.R
\name{fitting_prep_2}
\alias{fitting_prep_2}
\title{rDolphin: fitting_prep_2}
\usage{
fitting_prep_2(
  Xdata,
  Ydata,
  initial_fit_parameters,
  program_parameters,
  created_baseline,
  max_shift,
  min_shift,
  max_intensity,
  min_intensity,
  max_width,
  min_width,
  spectrum_index,
  signal_index
)
}
\arguments{
\item{Xdata}{Dataframe of sample chemical shift (ppm) values.}

\item{Ydata}{Dataframe of intensity values.}

\item{initial_fit_parameters}{Dataframe of signal fitting parameters}

\item{program_parameters}{List of parameters relevant to fitting algorithm.}

\item{created_baseline}{Dataframe of baseline intensity values}

\item{max_shift}{Dataframe of max chemical shift values}

\item{min_shift}{Dataframe of min chemical shift values}

\item{max_intensity}{Dataframe of max intensity values}

\item{min_intensity}{Dataframe of min intensity values}

\item{max_width}{Dataframe of max half bandwidth values}

\item{min_width}{Dataframe of min half bandwidth values}

\item{spectrum_index}{Index of sample spectrum to be fitted}

\item{signal_index}{Index of signal to be fitted}
}
\description{
rDolphin: fitting_prep_2
}
\details{
This function is based on code that was forked directly from rDolphin (https://github.com/danielcanueto/rDolphin/tree/master/R). This is function is used to collate
the data necessary for the downstream fitting algorithm and associated functions, assuming that profiling has occurred at least once before.
}
