% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rD_fittingloop.R
\name{fittingloop}
\alias{fittingloop}
\title{rDolphin: fittingloop}
\usage{
fittingloop(FeaturesMatrix, Xdata, Ydata, program_parameters)
}
\arguments{
\item{FeaturesMatrix}{Dataframe containing signal parameter information (e.g. multiplicity, roof effects)}

\item{Xdata}{Dataframe of sample chemical shift (ppm) values.}

\item{Ydata}{Dataframe of intensity values.}

\item{program_parameters}{List of parameters relevant to fitting algorithm.}
}
\description{
rDolphin: fittingloop
}
\details{
This function is based on code that was forked directly from rDolphin (https://github.com/danielcanueto/rDolphin/tree/master/R). This is the function
responsible for optimizing the fit of the signals of interest. Modifications have been made to the original function to allow for
the fitting of double-doublets. Generally speaking, this involves the addition of new fields (a second J-coupling field and second
roof effect field) containing parameters necessary to fit this new pattern.
}
