% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_reference_data.R
\name{get_spectra_data}
\alias{get_spectra_data}
\title{Pulls spectra character data for a given entry id from a locally maintained respository of JSON formated BMRM-STAR files}
\usage{
get_spectra_data(ID_list)
}
\arguments{
\item{ID_list}{list of BMRB IDs; For metabolomics entries it is just a number with BMRB prefix (example: 'bmse000028')}
}
\value{
R data frame that contains  Spectral_transition_char data for a given entry ID
}
\description{
Pulls spectra character data for a given entry id from a locally maintained respository of JSON formated BMRM-STAR files
}
\examples{
# df <- get_spectra_data("path/to/json/", as.bmseList(list("56-41-7","2613-02-7"), 400, 'D2O'))
# Downloads data from BMRB metabolomics database
}
\seealso{
\code{\link{show_file}}
}
