% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_utils.R
\name{\%ni\%}
\alias{\%ni\%}
\title{Negation of \%in\% operator}
\usage{
a \%ni\% b
}
\arguments{
\item{a}{a vector}

\item{b}{a vector}
}
\value{
a logical-valued vector with length identical to vector "a".
}
\description{
Negation of \%in\% operator
}
\details{
This function determines which of the elements of "a" are not contained within "b"
}
