% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_reference_data.R
\name{nearest_match_metabs}
\alias{nearest_match_metabs}
\title{Return nearest match metabolites, sorted by Euclidean distance}
\usage{
nearest_match_metabs(roi_df, pH, instrument_strength)
}
\arguments{
\item{roi_df}{data.frame, output of \code{roi_ref_export} with parameter \code{return_all = TRUE}}

\item{pH}{single numeric value specifying experimental pH}

\item{instrument_strength}{single numeric value specifying experimental instrument strength}
}
\value{
data.frame
}
\description{
Return nearest match metabolites, sorted by Euclidean distance
}
