% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmRapp.R
\name{nmRapp}
\alias{nmRapp}
\title{Semi-Automated Targeted Metabolite NMR Profiling}
\usage{
nmRapp(...)
}
\arguments{
\item{...}{additional arguments to be passed to shinyApp().}
}
\description{
Semi-Automated Targeted Metabolite NMR Profiling
}
\details{
Calls to this function create a new instance of a shiny app developed for semi-automated 
targeted metabolite NMR profiling.
}
