% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rD_output_generator.R
\name{output_generator}
\alias{output_generator}
\title{rDolphin: output_generator}
\usage{
output_generator(
  signals_to_quantify,
  fitted_signals,
  Ydata,
  Xdata,
  signals_parameters,
  multiplicities,
  buck_step
)
}
\arguments{
\item{signals_to_quantify}{}

\item{fitted_signals}{}

\item{Ydata}{}

\item{Xdata}{}

\item{signals_parameters}{}

\item{multiplicities}{}

\item{buck_step}{}
}
\description{
rDolphin: output_generator
}
\details{
This function was forked directly from rDolphin (https://github.com/danielcanueto/rDolphin/tree/master/R). This function formats
the output generated by the fitting routine and computes error metrics based on this output.
}
