% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_plot_xpmt_data.R
\name{plot_xpmt_data}
\alias{plot_xpmt_data}
\title{Plot uploaded experimental data}
\usage{
plot_xpmt_data(xpmt_data, sourceid, sample_to_plot = NULL, brushed_data = NULL)
}
\arguments{
\item{xpmt_data}{A dataframe with at least two columns. See details.}

\item{sourceid}{A string specifying the id of the generated plotly figure.}

\item{sample_to_plot}{A string specifying denoting which sample should be plotted (for the main plot).}

\item{brushed_data}{plotly eventData corresponding to a brush event. This is used to inform subplot generation.}
}
\value{
A plotly object
}
\description{
Plot uploaded experimental data
}
\details{
This function generates a plotly figure of the user-uploaded experimental data. The first column of the dataframe
required by the argument "xpmt_data" should contain the experimental PPM values. Subsequent columns should contain the
corresponding intensity values for each experimental sample.
}
