% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_profiling.R
\name{profiling_completeviewUI}
\alias{profiling_completeviewUI}
\title{Module: UI element to display trelliscope plot of fitted spectra}
\usage{
profiling_completeviewUI(id)
}
\arguments{
\item{id}{A string denoting the namespace id.}
}
\description{
Module: UI element to display trelliscope plot of fitted spectra
}
\details{
This is one of the UI components for the module created to handle all functions related to profiling
of experimental data, based on user edits to the fitting parameters of a pre-specified list of
reference (target) metabolites. The value provided for 'id' should be identical across the following:
profiling_controlsUI(), profiling_trelliscopeUI(), profiling_quantificationUI(), profiling_summaryUI(), and
profilingServer().

This module component provides the UI element that allow users to:
1) View (within a trelliscope display) the fitted spectra over the sample spectra for all samples and target metabolites
}
