% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rD_profiling_func.R
\name{profiling_func}
\alias{profiling_func}
\title{rDolphin: profiling_func}
\usage{
profiling_func(
  spectrum_index,
  signals_codes,
  imported_data,
  ROI_buckets,
  fitting_type,
  program_parameters,
  Xdata,
  Ydata,
  final_output,
  reproducibility_data,
  ROI_profile,
  baselinedataset,
  signals_to_quantify,
  pb = NULL,
  reimplementation = F,
  max_shift = NULL,
  min_shift = NULL,
  max_intensity = NULL,
  min_intensity = NULL,
  max_width = NULL,
  min_width = NULL,
  signal_index = NULL
)
}
\arguments{
\item{spectrum_index}{}

\item{signals_codes}{}

\item{imported_data}{}

\item{ROI_buckets}{}

\item{fitting_type}{}

\item{program_parameters}{}

\item{Xdata}{}

\item{Ydata}{}

\item{final_output}{}

\item{reproducibility_data}{}

\item{ROI_profile}{}

\item{baselinedataset}{}

\item{signals_to_quantify}{}

\item{pb}{}

\item{reimplementation}{}

\item{max_shift}{}

\item{min_shift}{}

\item{max_intensity}{}

\item{min_intensity}{}

\item{max_width}{}

\item{min_width}{}

\item{signal_index}{}
}
\description{
rDolphin: profiling_func
}
\details{
This function was forked directly from rDolphin (https://github.com/danielcanueto/rDolphin/tree/master/R). This is one of the
main functions used by rDolphin to perform automatic profiling. Modifications have been made to the original function to allow for
the fitting of double-doublets. Generally speaking, this involves the addition of new fields (a second J-coupling field and second
roof effect field) containing parameters necessary to fit this new pattern.
}
