% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_profiling.R
\name{profilingServer}
\alias{profilingServer}
\title{Module: Server functions specific to metabolite profiling and profiling result generation}
\usage{
profilingServer(id, xpmt_data, ref_data)
}
\arguments{
\item{id}{A string denoting the namespace id.}

\item{xpmt_data}{A reactive object containing experimental NMR data and associated metadata.}

\item{ref_data}{A reactive object containing target metabolite data, after having been edited.}
}
\value{
A reactive object containing a list of two elements. The first element of this list, final_output, is the
output of rDolphin's profiling procedure. The second element of the list, reproducibility_data, is some other component
generated by rDolphin's profiling procedure. For details, refer to rDolhpin's documentation and/or git repo.
}
\description{
Module: Server functions specific to metabolite profiling and profiling result generation
}
\details{
This is the server component for the module created to handle all functions related to profiling
of experimental data, based on user edits to the fitting parameters of a pre-specified list of
reference (target) metabolites. The value provided for 'id' should be identical across the following:
profiling_controlsUI(), profiling_trelliscopeUI(), profiling_quantificationUI(), profiling_summaryUI(), and
profilingServer().

This module component provides the back-end code that:
1) Performs the profiling for all target metabolites for all sample spectra
2) Generates the trelliscope plot and variety of tables used to visualize and summarize the profiling results
}
