% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_module_ref_data_modify.R
\name{ref_data_add_delUI}
\alias{ref_data_add_delUI}
\title{Module: UI elements specific to adding/removing target metabolites}
\usage{
ref_data_add_delUI(id)
}
\arguments{
\item{id}{A string denoting the namespace id.}
}
\description{
Copyright (C) 2022 Battelle Memorial Institute

 This program is free software; you can redistribute it and/or modify
 it under the terms of the GNU General Public License as published by
 the Free Software Foundation; either version 2 of the License, or
 (at your option) any later version.

 This program is distributed in the hope that it will be useful,
 but WITHOUT ANY WARRANTY; without even the implied warranty of
 MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
 GNU General Public License for more details.

 You should have received a copy of the GNU General Public License along
 with this program; if not, write to the Free Software Foundation, Inc.,
 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
}
\details{
This is one of the UI components for the module created to handle all editing of reference (target) metabolite data.
The value provided for 'id' should be identical across the following: ref_data_ROIeditingUI(), ref_data_add_delUI(),
ref_data_quantTab(), ref_data_profileUI(), ref_data_global_fittingparamsUI(), and ref_data_editingServer().

This module component provides the UI elements that allow users to:
1) Specify metabolite(s) to add or remove
2) Add/remove specified metabolite(s)

This module component also displays our internal reference database within a searchable, filterable table.
}
