% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_module_ref_data_modify.R
\name{ref_data_editingServer}
\alias{ref_data_editingServer}
\title{Module: Server functions specific to target metabolite data modification}
\usage{
ref_data_editingServer(id, xpmt_data, ref_data, ref_db)
}
\arguments{
\item{id}{A string denoting the namespace id.}

\item{xpmt_data}{A reactive object containing experimental NMR data and associated metadata.}

\item{ref_data}{A reactive object containing target metabolite data.}

\item{ref_db}{Dataframe of internal database of reference metabolite data.}
}
\value{
A reactive object containing the edited target (reference) metabolite data, with attributes describing the history of
changes made to the initial data; a logical value (TRUE/FALSE) indicating whether the user confirmed proceeding with profiling;
and a list of global profiling parameters.
}
\description{
Copyright (C) 2022 Battelle Memorial Institute

 This program is free software; you can redistribute it and/or modify
 it under the terms of the GNU General Public License as published by
 the Free Software Foundation; either version 2 of the License, or
 (at your option) any later version.

 This program is distributed in the hope that it will be useful,
 but WITHOUT ANY WARRANTY; without even the implied warranty of
 MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
 GNU General Public License for more details.

 You should have received a copy of the GNU General Public License along
 with this program; if not, write to the Free Software Foundation, Inc.,
 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
}
\details{
This is the server component for the module created to handle all editing of reference (target) metabolite data.
The value provided for 'id' should be identical across the following: ref_data_ROIeditingUI(), ref_data_add_delUI(),
ref_data_quantTab(), ref_data_profileUI(), and ref_data_editingServer().

This module component provides the back-end code that:
1) Populates the list of editable target metabolites that users may select from
2) Adds and/or removes target metabolites
3) Creates/formats reference (target) metabolite data for display and editing
4) Creates the dropdown containing several plot options.
5) Saves/reverts changes made to target metabolite data
6) Allows for plot-table interactivity such that changes made to certain plot elements reflect in the accompanying table and vice-versa
7) Initiates profiling of target metabolites
}
