% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_ref_data_modify.R
\name{ref_data_quantTab}
\alias{ref_data_quantTab}
\title{Module: UI element displaying target metabolite(s) to be quantified}
\usage{
ref_data_quantTab(id)
}
\arguments{
\item{id}{A string denoting the namespace id.}
}
\description{
Module: UI element displaying target metabolite(s) to be quantified
}
\details{
This is one of the UI components for the module created to handle all editing of reference (target) metabolite data.
The value provided for 'id' should be identical across the following: ref_data_ROIeditingUI(), ref_data_add_delUI(), 
ref_data_quantTab(), ref_data_profileUI(), and ref_data_editingServer().

This module component displays the set of target metabolite data that will be used for profiling. These data reflect any
and all edits made to the target metabolite data and are displayed within a searchable, filterable table.
}
