% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_module_ref_data_modify.R
\name{ref_data_ROIeditingUI}
\alias{ref_data_ROIeditingUI}
\title{Module: UI elements specific to reference data ROI editing}
\usage{
ref_data_ROIeditingUI(id)
}
\arguments{
\item{id}{A string denoting the namespace id.}
}
\description{
Copyright (C) 2022 Battelle Memorial Institute

 This program is free software; you can redistribute it and/or modify
 it under the terms of the GNU General Public License as published by
 the Free Software Foundation; either version 2 of the License, or
 (at your option) any later version.

 This program is distributed in the hope that it will be useful,
 but WITHOUT ANY WARRANTY; without even the implied warranty of
 MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
 GNU General Public License for more details.

 You should have received a copy of the GNU General Public License along
 with this program; if not, write to the Free Software Foundation, Inc.,
 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
}
\details{
This is one of the UI components for the module created to handle all editing of reference (target) metabolite data.
The value provided for 'id' should be identical across the following: ref_data_ROIeditingUI(), ref_data_add_delUI(),
ref_data_quantTab(), ref_data_profileUI(), and ref_data_editingServer().

This module component provides the UI elements that allow users to:
1) Select from a dropdown of several options that relate to the main plot. These options include the choice of
   sample spectrum to plot, the choice of target metabolites that they wish to edit and display over the sample spectrum,
   and whether subplots should be displayed on region select. The dropdown also indicates
   what filters (if any) have been applied to the experimental spectral data.
2) Save changes, revert the last saved changes, and revert all saved changes made to a given target
   metabolite's data (i.e. fitting parameters)
3) Check the fit of metabolites and make changes to global fitting parameters

This module component also displays target metabolite data within an interactive plot and table.
}
