% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_module_ref_data_upload.R
\name{ref_data_uploadServer}
\alias{ref_data_uploadServer}
\title{Module: Server functions specific to target metabolite data uploading}
\usage{
ref_data_uploadServer(id, xpmt_data, ref_db)
}
\arguments{
\item{id}{A string denoting the namespace id.}

\item{xpmt_data}{A reactive object containing experimental NMR data and associated metadata.}

\item{ref_db}{Dataframe of internal database of reference metabolite data.}
}
\value{
A reactive object with two elements. The first, $ref_data, is a dataframe containing the uploaded or manually specified
reference data, formatted as per roi_ref_export(). The second, $casno_not_in_db, is a list containing any CAS number(s)
and corresponding uploaded data entries that were not matched to entries within our internal database. This element will be null
if reference metabolites were manually specified or all uploaded reference metabolites were matched to internal database entries.
}
\description{
Copyright (C) 2022 Battelle Memorial Institute

 This program is free software; you can redistribute it and/or modify
 it under the terms of the GNU General Public License as published by
 the Free Software Foundation; either version 2 of the License, or
 (at your option) any later version.

 This program is distributed in the hope that it will be useful,
 but WITHOUT ANY WARRANTY; without even the implied warranty of
 MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
 GNU General Public License for more details.

 You should have received a copy of the GNU General Public License along
 with this program; if not, write to the Free Software Foundation, Inc.,
 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
}
\details{
This is the server component for the module created to handle all uploading/specification of reference (target) metabolite data.
The value provided for 'id' should be identical across ref_data_uploadUI() and ref_data_uploadServer().

This module component provides the back-end code that:
1) Dynamically toggles between the option sets for file upload or manual specification of reference metabolites
2) Reads in uploaded files and searches for columns containing CAS numbers based on suppplied column names
3) Processes and formats uploaded or specified data via nmRanalysis
4) Outputs a warning message if uploaded data are not matched to internal reference data
}
