% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_ref_data_upload.R
\name{ref_data_uploadUI}
\alias{ref_data_uploadUI}
\title{Module: UI elements specific to reference metabolite data uploading}
\usage{
ref_data_uploadUI(id, ref_db)
}
\arguments{
\item{id}{A string denoting the namespace id.}

\item{ref_db}{Dataframe of internal database of reference metabolite data.}
}
\description{
Module: UI elements specific to reference metabolite data uploading
}
\details{
This is the UI component for the module created to handle all uploading/specification of reference (target) metabolite data.
The value provided for 'id' should be identical across ref_data_uploadUI() and ref_data_uploadServer().

This module component provides the UI elements that allow users to:
1) Select an import method for the target metabolite data. Users may choose to either upload a .xlsx file
containing a column of CAS numbers or users may choose to manually specify metabolites by name. Note that
autocompletion functionality is available for manual specification.
2) Import target (reference) metabolite data via the selected import method.
}
