% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_refmet_save_update.R
\name{refmet_save_update}
\alias{refmet_save_update}
\title{Save updates to reference metabolite fitting parameters}
\usage{
refmet_save_update(updated_refmet, rvlist)
}
\arguments{
\item{updated_refmet}{A string providing the name of the edited target (reference) metabolite.}

\item{rvlist}{A list. See details.

The object supplied to "rvlist" is technically a reactive values object, but these objects behave very similarly to lists.
The elements of rvlist are as follows:
1) change_counter: a named list where each element is a counter for the number of changes applied to the particular target metabolite.
2) refchanges: A named list of lists, where each element is a list documenting the changes saved for the particular target metabolite.
3) unsaved_change: A named list where each element contains any changed data which has yet to be saved for a particular target metabolite.
4) user_reference_data: The initially uploaded set of target metabolite data containing all specified target metabolites and associated fitting parameter information.
5) dspedt_user_reference_data: The fitting parameter information specified to a particular target metabolite.
6) dspedt_user_xpmt_data: Experimental data filtered to include only PPM values in the neighborhood of the peak centers of the target metabolite featured in dspedt_user_reference_data}
}
\value{
A list containing the same elements as rvlist, but with values updated accordingly.
}
\description{
Copyright (C) 2022 Battelle Memorial Institute

 This program is free software; you can redistribute it and/or modify
 it under the terms of the GNU General Public License as published by
 the Free Software Foundation; either version 2 of the License, or
 (at your option) any later version.

 This program is distributed in the hope that it will be useful,
 but WITHOUT ANY WARRANTY; without even the implied warranty of
 MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
 GNU General Public License for more details.

 You should have received a copy of the GNU General Public License along
 with this program; if not, write to the Free Software Foundation, Inc.,
 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
}
