% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{refmets}
\alias{refmets}
\title{Internal database of reference metabolites}
\format{
A data frame with 2224 rows and 6 variables:
\describe{
  \item{Entry_ID}{BMSE ID number}
  \item{CASno}{CAS number}
  \item{Field_strength}{Instrument field strength}
  \item{Solute}{The name of the measured metabolite}
  \item{Solvent}{Solvent used for experiment}
  \item{pH}{pH}
}
}
\source{
TBW
}
\usage{
refmets
}
\description{
A dataset containing the fitting parameter information for 2224 entries of several metabolites measured
under various experimental conditions.
}
\keyword{datasets}
