% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_ROI_line_gen.R
\name{ROI_line_gen}
\alias{ROI_line_gen}
\title{Generate plotly shapes (lines) representative of regions of interest for a specific target metabolite}
\usage{
ROI_line_gen(data)
}
\arguments{
\item{data}{A dataframe containing the fitting parameter information. See details.}
}
\value{
A list containing information necessary to create plotly shapes (lines).
}
\description{
Generate plotly shapes (lines) representative of regions of interest for a specific target metabolite
}
\details{
The dataframe required by data should contain all of the fitting parameter information for a set of metabolite(s).
 This dataframe should have the following named columns, in the following order: "ROI left edge (ppm)",
"ROI right edge (ppm)", "Quantification Mode", "Metabolite", "Quantification Signal", "Chemical shift(ppm)",
"Chemical shift tolerance (ppm)", "Half bandwidth (Hz)", "Multiplicity", "J coupling (Hz)", "Roof effect",
"Quantify", "HMDB_code", and "rowid".
}
