% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_rDolphin.R
\name{run_rDolphin}
\alias{run_rDolphin}
\title{Wrapper for running rDolphin}
\usage{
run_rDolphin(
  imported_data,
  dir = NULL,
  optimization = TRUE,
  spectra_to_profile = NULL
)
}
\arguments{
\item{imported_data}{an rDolphin data object, often created by \code{ppmData_to_rDolphin}}

\item{dir}{Directory where to save results, default is NA, which will create a folder in your working directory.}

\item{optimization}{Logical, should profiling quality be maximized through analysis of signal parameters. Default is TRUE.}

\item{spectra_to_profile}{A vector of spectrum indices. Defaults to NULL, profiling all spectra.}
}
\description{
Copyright (C) 2022 Battelle Memorial Institute

 This program is free software; you can redistribute it and/or modify
 it under the terms of the GNU General Public License as published by
 the Free Software Foundation; either version 2 of the License, or
 (at your option) any later version.

 This program is distributed in the hope that it will be useful,
 but WITHOUT ANY WARRANTY; without even the implied warranty of
 MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
 GNU General Public License for more details.

 You should have received a copy of the GNU General Public License along
 with this program; if not, write to the Free Software Foundation, Inc.,
 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.

Runs rDolphin using imported_data (the output from \code{\link{ppmData_to_rDolphin}}) and \code{\link{opt_rDolphin}}
}
\author{
Allison Thompson
}
