% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rD_save_output.R
\name{save_output}
\alias{save_output}
\title{rDolphin: save_output}
\usage{
save_output(
  spectrum_index,
  signals_codes,
  results_to_save,
  buck_step,
  final_output
)
}
\arguments{
\item{spectrum_index}{}

\item{signals_codes}{}

\item{results_to_save}{}

\item{buck_step}{}

\item{final_output}{}
}
\description{
rDolphin: save_output
}
\details{
This function was forked directly from rDolphin (https://github.com/danielcanueto/rDolphin/tree/master/R). This function saves
the formatted output of the quantification generated by the automatic profiling routine.
}
