% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_reference_data.R
\name{show_file}
\alias{show_file}
\title{Imports whole STAR file for a given entry id from a locally maintained respository of JSON formated BMRM-STAR files}
\usage{
show_file(ID)
}
\arguments{
\item{ID}{single BMRB ID; For metabolomics entries it is just a number with bmrb prefix (example: 'bmse000028');}
}
\value{
R list that contains STAR formatted data for a given entry ID
}
\description{
Imports whole STAR file for a given entry id from a locally maintained respository of JSON formated BMRM-STAR files
}
\examples{
# file <- show_file("path/to/json/","bmse000028")
# Downloads entire STAR file from BMRB metabolomics database
}
\seealso{
\code{\link{get_spectra_data}}
}
