% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rD_signal_fitting.R
\name{signal_fitting}
\alias{signal_fitting}
\title{rDolphin: signal_fitting}
\usage{
signal_fitting(parS, Xdata, multiplicities, roof_effect, roof_effect2, freq)
}
\arguments{
\item{parS}{vector of signal fitting parameters}

\item{Xdata}{Dataframe of sample chemical shift (ppm) values.}

\item{multiplicities}{vector of signal multiplicity values}

\item{roof_effect}{vector of signal roof effect values}

\item{roof_effect2}{vector of signal roof effect 2 values. Applicable only if one or more signals is of non-first-order multiplicity.}

\item{freq}{spectrometer frequency}
}
\description{
rDolphin: signal_fitting
}
\details{
This function is based on code that was forked directly from rDolphin (https://github.com/danielcanueto/rDolphin/tree/master/R). This is the function
that fits the lineshape patterns according to the specified fitting parameters. Modifications have been made to the original function to allow for
the fitting of double-doublets. Generally speaking, this involves the addition of new fields (a second J-coupling field and second
roof effect field) containing parameters necessary to fit this new pattern and new fitting equations that make use of these additional
fields.
}
