% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_module_xpmt_data_upload.R
\name{xpmt_data_uploadServer}
\alias{xpmt_data_uploadServer}
\title{Module: Server functions specific to experimental data uploading}
\usage{
xpmt_data_uploadServer(id)
}
\arguments{
\item{id}{A string denoting the namespace id.}
}
\value{
A reactive object containing the uploaded experimental data and parameters saved as an nmRanalysis ppmData object.
}
\description{
Copyright (C) 2022 Battelle Memorial Institute

 This program is free software; you can redistribute it and/or modify
 it under the terms of the GNU General Public License as published by
 the Free Software Foundation; either version 2 of the License, or
 (at your option) any later version.

 This program is distributed in the hope that it will be useful,
 but WITHOUT ANY WARRANTY; without even the implied warranty of
 MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
 GNU General Public License for more details.

 You should have received a copy of the GNU General Public License along
 with this program; if not, write to the Free Software Foundation, Inc.,
 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
}
\details{
This is the UI component for the module created to handle the uploading of experimental data.
The value provided for 'id' should be identical across xpmt_data_uploadUI() and xpmt_data_uploadServer().

This module component provides the back-end code that:
1) Processes uploaded data based on the provided experimental conditions. Data are saved as an nmRanalysis ppmData object.
}
