% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_xpmt_data_upload.R
\name{xpmt_data_uploadUI}
\alias{xpmt_data_uploadUI}
\title{Module: UI elements specific to experimental data uploading}
\usage{
xpmt_data_uploadUI(id, ref_db)
}
\arguments{
\item{id}{A string denoting the namespace id.}

\item{ref_db}{Dataframe of internal database of reference metabolite data.}
}
\description{
Module: UI elements specific to experimental data uploading
}
\details{
This is the UI component for the module created to handle the uploading of experimental data.
The value provided for 'id' should be identical across xpmt_data_uploadUI() and xpmt_data_uploadServer().

This module component provides the UI elements that allow users to:
1) Upload the .csv file containing the experimental PPM and Intensity data for all samples. The first column of
this datafile should provide the PPM values and subsequent columns the corresponding intensity values for each
sample.
2) Upload the .csv file containing experimental metadata. Currently, the only metadata processed are sample labels.
3) Specify experimental pH, instrument strength, and the solvent used. The available choices for each of these
"experimental conditions" is constrained by the conditions available in our internal database.
}
