% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_xpmt_data_viz.R
\name{xpmt_data_vizServer}
\alias{xpmt_data_vizServer}
\title{Module: Server functions specific to experimental data and metadata visualization}
\usage{
xpmt_data_vizServer(id, xpmt_data)
}
\arguments{
\item{id}{A string denoting the namespace id.}

\item{xpmt_data}{The reactive object returned by the xpmt_data_uploadServer() server module containing
the experimental data as a nmRanalysis ppmData object.}
}
\value{
A reactive object containing the uploaded experimental data with any modifications (i.e. filters) applied. Note that
any applied filters are stored as an attribute of the object. The attribute contains both the filtered range and the set of data
that was filtered within this range.
}
\description{
Module: Server functions specific to experimental data and metadata visualization
}
\details{
This is the UI component for the module created to handle the visualization of experimental data and metadata.
The value provided for 'id' should be identical across xpmt_data_vizUI(), xpmt_data_vizoptionsUI(),
xpmt_metadata_vizUI, and xpmt_data_vizServer().

This module component provides the back-end code that:
1) Generates the plotly figure of experimental data
2) Generates the datatables of experimental data and metadata
3) Creates and displays subplots corresponding to selected regions
4) Applies or removes filters to the data
}
