% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_xpmt_data_viz.R
\name{xpmt_metadata_vizUI}
\alias{xpmt_metadata_vizUI}
\title{Module: UI elements specific to experimental metadata visualization}
\usage{
xpmt_metadata_vizUI(id)
}
\arguments{
\item{id}{A string denoting the namespace id.}
}
\description{
Module: UI elements specific to experimental metadata visualization
}
\details{
This is the UI component for the module created to handle the visualization of experimental metadata.
The value provided for 'id' should be identical across xpmt_data_vizUI(), xpmt_data_vizoptionsUI(),
xpmt_metadata_vizUI, and xpmt_data_vizServer().

This module component provides the UI element that allows users to:
1) Visualize uploaded experimental metadata through a searchable datatable.
}
